feat: nft staking

+++ nft-staking/programs/nft-staking/src/state/stake_account.rs
@@ -6,5 +6,5 @@ pub struct StakeAccount {
     pub owner: Pubkey,
     pub mint: Pubkey,
     pub staked_at: i64,
-    pub bump: u64,
+    pub bump: u8,
 }
